requirements.txt
torch
biopython
einops
matplotlib
pandas
scipy
scikit-learn
tqdm
wandb
ipykernel
easydict
ipdb
tqdm
submitit
dm-tree
fair-esm
